refactor c++ decompression stuff out of AtexAsm
